refactor: code clean up
